Added CRON support
It is not necessary to authenticate users if this script is run via cron. Therefore, it will now only authenticate if output to the browser is enabled. If the user is not authenticated, it will continue to generate the script while not outputting to the browser, updating any other output methods enabled.